chore(team): split mailbox store delivery
